refactorizado

+++ app/src/main/java/com/example/todotasks/domain/model/SubTask.kt
@@ -2,6 +2,6 @@ package com.example.todotasks.domain.model
 
 data class SubTask(
     val id: Long = 0,
-    val idTask: Int,
+    val idTask: Long,
     val title: String,
     val completed: Boolean = false)
